Support for pruned scan for required fields.
https://github.com/databricks/spark-xml/issues/21
This library will read all data but does not have to check all the values in rows and convert (cast) them all.
This PR adds the feature for pruned scan for this.

For the reason similar with https://github.com/databricks/spark-csv/issues/219, this should parse all when dropping rows. This PR is similar with https://github.com/databricks/spark-csv/pull/169.

I did not add a test for this as it looks existing tests covers this.

Author: hyukjinkwon <gurwls223@gmail.com>

Closes #46 from HyukjinKwon/ISSUE-21-pruned.