Handling inconsistent element attributes
#93 #94

The size of the nested row for that case itself was different with the schema. So, when it accesses to the fields, it emits `java.lang.ArrayIndexOutOfBoundsException` exception as described in the issue above.

This PR sets `null` explicitly for the fields as well like the other field so that there is no inconsistent fields.

Author: hyukjinkwon <gurwls223@gmail.com>

Closes #95 from HyukjinKwon/ISSUE-93-incon-attr.