Do not allow empty strings for attributePrefix, valueTag and rowTag
This PR adds requirements for `attributePrefix`, `valueTag` and `rowTag`.

Since those options are used to differentiate attributes, `attributePrefix` should not be just a empty string

`valueTag` is used for a column name so it should not be empty string.

Lastly, we should set the explicit range for each row via `rowTag`. Therefore, this should not be just empty string too.

Author: hyukjinkwon <gurwls223@gmail.com>

Closes #170 from HyukjinKwon/ISSUE-153.